requirements.txt
amqp==5.0.6
appdirs==1.4.4
apturl==0.5.2
asgiref==3.4.1
astroid==2.3.3
attrs==21.2.0
bcrypt==3.1.7
billiard==3.6.4.0
blinker==1.4
Brlapi==0.7.0
CacheControl==0.12.10
cached-property==1.5.2
cachetools==4.2.4
catfish==1.4.13
certifi==2019.11.28
chardet==3.0.4
Click==7.0
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.8
django-cors-headers==3.2.1
django-jalali-date==0.3.2
django-model-utils==4.2.0
django-widget-tweaks==1.4.9
djangorestframework==3.11.2
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
firebase-admin==5.1.0
flake8==3.7.9
future==0.18.2
google-api-core==2.2.2
google-api-python-client==2.30.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.41.1
grpcio-status==1.41.1
gunicorn==20.1.0
httplib2==0.20.2
idna==2.8
isodate==0.6.0
isort==4.3.21
jdatetime==3.6.4
keyring==18.0.1
kombu==5.1.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lightdm-gtk-greeter-settings==1.2.2
lockfile==0.12.2
louis==3.12.0
lxml==4.6.3
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mccabe==0.6.1
menulibre==2.2.1
monotonic==1.5
msgpack==1.0.2
mugshot==0.4.2
netifaces==0.10.4
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
packaging==21.2
pandas==1.3.4
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
powerline-status==2.7
prompt-toolkit==3.0.20
proto-plus==1.19.8
protobuf==3.19.1
psutil==5.5.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.5.0
pycups==1.9.73
pycurl==7.43.0.2
pyflakes==2.1.1
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.15
pylint-plugin-utils==0.6
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
PyQt5==5.14.1
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rsa==4.7.2
SecretStorage==2.3.1
sgt-launcher==0.2.5
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sqlparse==0.4.2
systemd-python==234
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.25.8
usb-creator==0.3.7
vine==5.0.0
wadllib==1.3.3
wcwidth==0.2.5
wrapt==1.11.2
xcffib==0.8.1
xkit==0.0.0
zeep==3.4.0
